Renamed shared/http-types => shared/api. And changed file names accordingly.
