+++ frontend/src/hooks/api/useGetRecipes.api.ts
@@ -1,4 +1,4 @@
-import { type GetAllRecipesRes } from "../../../../shared/http-types/recipe/getAllRecipes.http-type";
+import { type GetAllRecipesRes } from "../../../../shared/api/recipe/getAllRecipes.http-type";
 import { useQuery } from "@tanstack/react-query";
 import api from "../../api";
