+++ frontend/src/hooks/api/useSaveChef.api.ts
@@ -3,7 +3,7 @@ import { useQueryClient, useMutation } from "@tanstack/react-query";
 import type {
   SaveChefReq,
   SaveChefRes,
-} from "../../../../shared/http-types/chef/saveChef.api";
+} from "../../../../shared/api/chef/saveChef.api";
 import { USE_GET_CHEFS_KEY } from "./useGetChefs.api";
 import type { ChefDetails } from "@shared/types/chef.type";
