+++ frontend/src/hooks/api/useSaveRecipe.api.ts
@@ -3,7 +3,7 @@ import { useQueryClient, useMutation } from "@tanstack/react-query";
 import type {
   SaveRecipeReq,
   SaveRecipeRes,
-} from "../../../../shared/http-types/recipe/saveRecipe.http-type";
+} from "../../../../shared/api/recipe/saveRecipe.http-type";
 import type { RecipeDetails } from "../../../../shared/types/recipe.type";
 import { USE_GET_RECIPES_KEY } from "./useGetRecipes.api";
 import { USE_GET_RECIPE_BY_UUID_KEY } from "./useGetRecipeByUuid.api";
